third_party/blink/web_tests/fast/forms/reset-autofilled-expected.txt

This tests that a field's autofilled state is reverted when the form is reset. It can only be run using the test harness.

  
PASS successfullyParsed is true

TEST COMPLETE

PASS countAutofilledElements() is 3
PASS document.querySelectorAll(':autofill').length is 3
PASS countAutofilledElements() is 0
PASS document.querySelectorAll(':autofill').length is 0
